Add main logic of the Sidebar
Added:
---

* general logic of Sidebar state switching ("Open" | "Closed");

* additional logic for opening the Sidebar using the Search field:
when you click or enter something in the Search field;

* additional logic for opening the Sidebar when clicking on
the "Projects" button;

===

In addition, added logic to open the list of projects when
clicking on the "Projects" button.

===

Among other things, a separate utility module has been created,
in which all the constant and utilitarian values that are reused
in the project are placed.