Add logic of theme switch
Added a script describing the basic logic of theme switching.

At the moment, the script adds the desired class for <body>
when you click on the theme-switcher.

In addition to adding a class for <body>, the script also
makes changes to the state of the signature text:
if the «Dark» theme is active now, the inscription will
"Turn on the «Light» theme" and vice versa.
---

In addition to the theme switching logic, styles have also
been added for the appearance of the theme-switcher.